+++ src/components/landingPage/FeaturedComponentBase.tsx
